+++ src/block-kit/elements/ConversationsSelect.ts
@@ -15,6 +15,9 @@ import { InputComponentProps, wrapInInput } from '../layout/Input'
 import { BuiltInComponent, createComponent } from '../../jsx'
 import { DistributedProps, coerceToInteger } from '../../utils'
 
+declare const conversationIdString: unique symbol
+type ConversationIdString = string & { [conversationIdString]?: never }
+
 interface SingleConversationsSelectProps
   extends ActionProps,
     ConfirmableProps,
@@ -22,14 +25,26 @@ interface SingleConversationsSelectProps
     FilterProps {
   children?: never
 
-  /** A string of ID for the initially selected conversation. */
-  initialConversation?: string
+  /**
+   * A string of ID for the initially selected conversation. In multiple select,
+   * you can set multiple conversation IDs through array.
+   *
+   * `initialConversation` accepts the conversation id prefixed with `C`, `D`,
+   * and `G`. In addition, jsx-slack recognizes the special value `current` for
+   * indicating the origin conversation that the container surface belongs to.
+   *
+   * **NOTE**: `current` corresponds to `default_to_current_conversation` field
+   * in Slack API. It will be ignored if defined initial conversations, so
+   * multiple conversations select cannot specify `current` along with specific
+   * conversations.
+   */
+  initialConversation?: ConversationIdString | 'current'
 
   /** The placeholder text shown in select field. */
   placeholder?: string
 
   /** An alias into `initialConversation` prop. */
-  value?: string
+  value?: ConversationIdString | 'current'
 }
 
 interface MultiConversationsSelectProps
@@ -37,15 +52,21 @@ interface MultiConversationsSelectProps
     SingleConversationsSelectProps,
     'initialConversation' | 'value'
   > {
-  /**
-   * In multiple select, you can set multiple conversation IDs through array.
-   */
-  initialConversation?: string | string[]
-  value?: string | string[]
+  initialConversation?:
+    | ConversationIdString
+    | 'current'
+    | ConversationIdString[]
+
+  value?: ConversationIdString | 'current' | ConversationIdString[]
 }
 
 type ConversationsSelectElement = SlackConversationsSelect & {
+  default_to_current_conversation?: boolean
   response_url_enabled?: boolean
+}
+
+type MultiConversationsSelectElement = MultiConversationsSelect & {
+  default_to_current_conversation?: boolean
 }
 
 export type ConversationsSelectProps = DistributedProps<
@@ -66,24 +87,38 @@ export type ConversationsSelectProps = DistributedProps<
  */
 export const ConversationsSelect: BuiltInComponent<ConversationsSelectProps> = createComponent<
   ConversationsSelectProps,
-  ConversationsSelectElement | MultiConversationsSelect | InputBlock
+  ConversationsSelectElement | MultiConversationsSelectElement | InputBlock
 >('ConversationsSelect', (props) => {
   const action_id = props.actionId || props.name
   const filterComposition = filter(props)
   const placeholder =
     props.placeholder !== undefined ? plainText(props.placeholder) : undefined
 
-  return wrapInInput<ConversationsSelectElement | MultiConversationsSelect>(
+  const initialConversationsSet = new Set<string>(
+    ((v) => ([] as string[]).concat(v ?? []))(
+      props.initialConversation || props.value
+    )
+  )
+
+  const defaultToCurrentConversation =
+    initialConversationsSet.delete('current') || undefined
+
+  const initialConversations =
+    initialConversationsSet.size > 0
+      ? [...initialConversationsSet.values()]
+      : undefined
+
+  return wrapInInput<
+    ConversationsSelectElement | MultiConversationsSelectElement
+  >(
     props.multiple
       ? {
           type: 'multi_conversations_select',
           action_id,
           placeholder,
-          initial_conversations: ((v) =>
-            v !== undefined ? ([] as string[]).concat(v) : undefined)(
-            props.initialConversation || props.value
-          ),
+          initial_conversations: initialConversations,
           filter: filterComposition,
+          default_to_current_conversation: defaultToCurrentConversation,
           max_selected_items: coerceToInteger(props.maxSelectedItems),
           confirm: props.confirm as any,
         }
@@ -91,8 +126,9 @@ export const ConversationsSelect: BuiltInComponent<ConversationsSelectProps> = c
           type: 'conversations_select',
           action_id: props.actionId || props.name,
           placeholder,
-          initial_conversation: props.initialConversation || props.value,
+          initial_conversation: initialConversations?.[0],
           filter: filterComposition,
+          default_to_current_conversation: defaultToCurrentConversation,
           response_url_enabled:
             props.responseUrlEnabled !== undefined
               ? !!props.responseUrlEnabled